Imágenes y componentes arreglados
