+++ components/analisis/static/style.css
