+++ components/analisis/templates/layout_analisis.html
@@ -66,7 +66,6 @@
 			</div>
 
 		</div>
-		</div>
 	</footer>
 
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
