+++ components/analisis/templates/subir.html
@@ -2,7 +2,6 @@
 
 {% block content %}
 
-<div class="container top-space">
   <center>
     <h1>Análisis de imágenes</h1>
     <form enctype="multipart/form-data" method="post" action="{% url 'analisis' %}">
@@ -21,7 +20,6 @@
       <li><a class="btn btn-default" href="{% url 'resultados' %}">Analizar</a></li>
     </div>
   </center>
-</div>
 
 
 {% endblock %}
