+++ components/busqueda/static/style.css
