+++ components/experimento/views.py
@@ -3,9 +3,8 @@ from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
 from django.views.generic import ListView, DetailView
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
-from django.urls import reverse_lazy
 
-from components.home.forms import ExperimentoForm
+from components.experimento.forms import ExperimentoForm
 
 from components.experimento.models import Experimento
