+++ components/tratamiento/static/style.css
